AWS SDK 사용해서 메트릭 가져오고 모니터링, 관리자의 교육자 승인 로직 구현

+++ src/main/java/com/example/p2k/course/CourseService.java
@@ -1,10 +1,12 @@
 package com.example.p2k.course;
 
 import com.example.p2k._core.exception.Exception400;
+import com.example.p2k._core.exception.Exception403;
 import com.example.p2k._core.exception.Exception404;
 import com.example.p2k.courseuser.CourseUser;
 import com.example.p2k.courseuser.CourseUserRepository;
 import com.example.p2k.post.PostRepository;
+import com.example.p2k.user.Role;
 import com.example.p2k.user.User;
 import com.example.p2k.vm.Vm;
 import com.example.p2k.vm.VmRepository;
@@ -96,10 +98,11 @@ public class CourseService {
                 () -> new Exception404("해당 강좌를 찾을 수 없습니다.")
         );
 
-        if(course.getUser() == null){
+        if(course.getInstructorId() == null){
             throw new Exception400("해당 강좌의 교육자가 존재하지 않습니다.");
         }
-        List<Vm> vms = vmRepository.findUserIdAndCourseIdOpen(course.getUser().getId(), id);
+
+        List<Vm> vms = vmRepository.findUserIdAndCourseIdOpen(course.getInstructorId(), id);
         return new VmResponse.FindAllDTO(vms);
     }
 
@@ -112,7 +115,15 @@ public class CourseService {
     //강좌 생성
     @Transactional
     public void create(CourseRequest.SaveDTO requestDTO, User user){
-        Course course = Course.builder().name(requestDTO.getName()).description(requestDTO.getDescription()).build();
+        if(user.getPending() || user.getRole() != Role.ROLE_INSTRUCTOR){
+            throw new Exception403("권한이 없는 사용자입니다.");
+        }
+
+        Course course = Course.builder()
+                .name(requestDTO.getName())
+                .description(requestDTO.getDescription())
+                .instructorId(user.getId())
+                .build();
         courseRepository.save(course);
 
         CourseUser courseUser = CourseUser.builder().course(course).user(user).accept(true).build();
@@ -120,7 +131,10 @@ public class CourseService {
     }
 
     //수강생 관리
-    public CourseResponse.FindStudentsDTO findStudents(Long id){
+    public CourseResponse.FindStudentsDTO findStudents(Long id, User user){
+        if(user.getPending() || user.getRole() != Role.ROLE_INSTRUCTOR){
+            throw new Exception403("권한이 없는 사용자입니다.");
+        }
         List<User> students = courseUserRepository.findAcceptedUserByCourseId(id);
         return new CourseResponse.FindStudentsDTO(students);
     }
@@ -133,19 +147,28 @@ public class CourseService {
 
     //강좌 신청 수락
     @Transactional
-    public void accept(Long courseId, Long userId){
+    public void accept(Long courseId, Long userId, User user){
+        if(!user.getPending() || user.getRole() != Role.ROLE_INSTRUCTOR){
+            throw new Exception403("권한이 없는 사용자입니다.");
+        }
         courseUserRepository.updateAccept(courseId, userId);
     }
 
     //강좌 신청 거절
     @Transactional
-    public void reject(Long courseId, Long userId){
+    public void reject(Long courseId, Long userId, User user){
+        if(user.getPending() || user.getRole() != Role.ROLE_INSTRUCTOR){
+            throw new Exception403("권한이 없는 사용자입니다.");
+        }
         courseUserRepository.deleteByCourseIdAndUserId(courseId, userId);
     }
 
     //강좌 삭제
     @Transactional
-    public void delete(Long id){
+    public void delete(Long id, User user){
+        if(user.getPending() || user.getRole() != Role.ROLE_INSTRUCTOR){
+            throw new Exception403("권한이 없는 사용자입니다.");
+        }
         courseUserRepository.deleteByCourseId(id);
         postRepository.deleteAllByCourseId(id);
         courseRepository.deleteById(id);
